# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PerformanceOS AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
  organizationName: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface SessionUser {
  id: string;
  email: string;
  name: string;
  orgId: string;
  orgName: string;
  role: string;
}

export interface AuthResponse {
  user: SessionUser;
}

export interface KPIs {
  spend: number;
  revenue: number;
  roas: number;
  cpa: number;
  cpl: number;
  ctr: number;
  cpc: number;
  cpm: number;
  leads: number;
  conversions: number;
}

export interface CampaignIssue {
  id: string;
  name: string;
  platform: string;
  spend: number;
  roas: number;
  issue: string;
}

export interface CampaignOpportunity {
  id: string;
  name: string;
  platform: string;
  roas: number;
  conversions: number;
  suggestion: string;
}

export interface CampaignRow {
  id: string;
  name: string;
  platform: string;
  spend: number;
  revenue: number;
  roas: number;
  conversions: number;
  clicks: number;
  impressions: number;
}

export interface DashboardSummary {
  kpis: KPIs;
  needsAttention: CampaignIssue[];
  opportunities: CampaignOpportunity[];
  topPerformers: CampaignRow[];
  executiveSummary: string;
  aiRecommendations: string[];
}

export interface TimeSeriesPoint {
  date: string;
  spend: number;
  revenue: number;
  roas: number;
}

export interface PlatformSummary {
  platform: string;
  spend: number;
  revenue: number;
  roas: number;
  conversions: number;
}

export interface AnalyticsData {
  timeSeries: TimeSeriesPoint[];
  campaigns: CampaignRow[];
  platformsSummary: PlatformSummary[];
}

export interface Alert {
  id: string;
  type: string;
  severity: string;
  message: string;
  isResolved: boolean;
  createdAt: string;
}

export interface SyncLog {
  id: string;
  status: string;
  /** @nullable */
  errorMessage?: string | null;
  rowsSynced: number;
  timestamp: string;
}

export interface WindsorConnectionInfo {
  connected?: boolean;
  /** @nullable */
  syncStatus?: string | null;
  /** @nullable */
  lastSyncAt?: string | null;
  /** @nullable */
  createdAt?: string | null;
  syncLogs?: SyncLog[];
}

export interface WindsorInput {
  apiKey: string;
}

export interface SyncResult {
  rowsSynced: number;
}

export interface Conversation {
  id: string;
  title: string;
  createdAt: string;
}

export interface ConversationInput {
  title?: string;
}

export interface Message {
  id: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}