# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PerformanceOS AI API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: dashboard
    description: Dashboard metrics
  - name: analytics
    description: Analytics data
  - name: alerts
    description: Performance alerts
  - name: integrations
    description: Windsor.ai integrations
  - name: athena
    description: Athena AI conversations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user and organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Unauthorized

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Unauthorized

  /analytics/data:
    get:
      operationId: getAnalyticsData
      tags: [analytics]
      summary: Get analytics data for charts and campaigns
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsData"
        "401":
          description: Unauthorized

  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: Get all alerts for the org
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
        "401":
          description: Unauthorized

  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Mark alert as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Not found

  /integrations/windsor:
    get:
      operationId: getWindsorConnection
      tags: [integrations]
      summary: Get Windsor.ai connection
      responses:
        "200":
          description: Connection info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WindsorConnectionInfo"
    post:
      operationId: connectWindsor
      tags: [integrations]
      summary: Connect Windsor.ai
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WindsorInput"
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WindsorConnectionInfo"
    delete:
      operationId: disconnectWindsor
      tags: [integrations]
      summary: Disconnect Windsor.ai
      responses:
        "200":
          description: Disconnected

  /integrations/windsor/sync:
    post:
      operationId: triggerSync
      tags: [integrations]
      summary: Trigger manual sync
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"

  /athena/conversations:
    get:
      operationId: getConversations
      tags: [athena]
      summary: Get all conversations
      responses:
        "200":
          description: Conversations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [athena]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /athena/conversations/{id}/messages:
    get:
      operationId: getConversationMessages
      tags: [athena]
      summary: Get messages for a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [athena]
      summary: Send a message to Athena
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegisterInput:
      type: object
      required: [email, password, name, organizationName]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        organizationName:
          type: string

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/SessionUser"

    SessionUser:
      type: object
      required: [id, email, name, orgId, orgName, role]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        orgId:
          type: string
        orgName:
          type: string
        role:
          type: string

    DashboardSummary:
      type: object
      required: [kpis, needsAttention, opportunities, topPerformers, executiveSummary, aiRecommendations]
      properties:
        kpis:
          $ref: "#/components/schemas/KPIs"
        needsAttention:
          type: array
          items:
            $ref: "#/components/schemas/CampaignIssue"
        opportunities:
          type: array
          items:
            $ref: "#/components/schemas/CampaignOpportunity"
        topPerformers:
          type: array
          items:
            $ref: "#/components/schemas/CampaignRow"
        executiveSummary:
          type: string
        aiRecommendations:
          type: array
          items:
            type: string

    KPIs:
      type: object
      required: [spend, revenue, roas, cpa, cpl, ctr, cpc, cpm, leads, conversions]
      properties:
        spend:
          type: number
        revenue:
          type: number
        roas:
          type: number
        cpa:
          type: number
        cpl:
          type: number
        ctr:
          type: number
        cpc:
          type: number
        cpm:
          type: number
        leads:
          type: number
        conversions:
          type: number

    CampaignIssue:
      type: object
      required: [id, name, platform, spend, roas, issue]
      properties:
        id:
          type: string
        name:
          type: string
        platform:
          type: string
        spend:
          type: number
        roas:
          type: number
        issue:
          type: string

    CampaignOpportunity:
      type: object
      required: [id, name, platform, roas, conversions, suggestion]
      properties:
        id:
          type: string
        name:
          type: string
        platform:
          type: string
        roas:
          type: number
        conversions:
          type: number
        suggestion:
          type: string

    CampaignRow:
      type: object
      required: [id, name, platform, spend, revenue, roas, conversions, clicks, impressions]
      properties:
        id:
          type: string
        name:
          type: string
        platform:
          type: string
        spend:
          type: number
        revenue:
          type: number
        roas:
          type: number
        conversions:
          type: number
        clicks:
          type: number
        impressions:
          type: number

    AnalyticsData:
      type: object
      required: [timeSeries, campaigns, platformsSummary]
      properties:
        timeSeries:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        campaigns:
          type: array
          items:
            $ref: "#/components/schemas/CampaignRow"
        platformsSummary:
          type: array
          items:
            $ref: "#/components/schemas/PlatformSummary"

    TimeSeriesPoint:
      type: object
      required: [date, spend, revenue, roas]
      properties:
        date:
          type: string
        spend:
          type: number
        revenue:
          type: number
        roas:
          type: number

    PlatformSummary:
      type: object
      required: [platform, spend, revenue, roas, conversions]
      properties:
        platform:
          type: string
        spend:
          type: number
        revenue:
          type: number
        roas:
          type: number
        conversions:
          type: number

    Alert:
      type: object
      required: [id, type, severity, message, isResolved, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
        severity:
          type: string
        message:
          type: string
        isResolved:
          type: boolean
        createdAt:
          type: string

    WindsorConnectionInfo:
      type: object
      properties:
        connected:
          type: boolean
        syncStatus:
          type: ["string", "null"]
        lastSyncAt:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
        syncLogs:
          type: array
          items:
            $ref: "#/components/schemas/SyncLog"

    WindsorInput:
      type: object
      required: [apiKey]
      properties:
        apiKey:
          type: string

    SyncResult:
      type: object
      required: [rowsSynced]
      properties:
        rowsSynced:
          type: integer

    SyncLog:
      type: object
      required: [id, status, rowsSynced, timestamp]
      properties:
        id:
          type: string
        status:
          type: string
        errorMessage:
          type: ["string", "null"]
        rowsSynced:
          type: integer
        timestamp:
          type: string

    Conversation:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string

    ConversationInput:
      type: object
      properties:
        title:
          type: string

    Message:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: string
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string

    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string