# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PerformanceOS AI API
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: string;
  type: string;
  severity: string;
  message: string;
  isResolved: boolean;
  createdAt: string;
}