# lib/api-zod/src/generated/types/analyticsData.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PerformanceOS AI API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignRow } from './campaignRow';
import type { PlatformSummary } from './platformSummary';
import type { TimeSeriesPoint } from './timeSeriesPoint';

export interface AnalyticsData {
  timeSeries: TimeSeriesPoint[];
  campaigns: CampaignRow[];
  platformsSummary: PlatformSummary[];
}