# lib/api-zod/src/generated/types/campaignIssue.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PerformanceOS AI API
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignIssue {
  id: string;
  name: string;
  platform: string;
  spend: number;
  roas: number;
  issue: string;
}