# lib/api-zod/src/generated/types/campaignOpportunity.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PerformanceOS AI API
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignOpportunity {
  id: string;
  name: string;
  platform: string;
  roas: number;
  conversions: number;
  suggestion: string;
}