# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PerformanceOS AI API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignIssue } from './campaignIssue';
import type { CampaignOpportunity } from './campaignOpportunity';
import type { CampaignRow } from './campaignRow';
import type { KPIs } from './kPIs';

export interface DashboardSummary {
  kpis: KPIs;
  needsAttention: CampaignIssue[];
  opportunities: CampaignOpportunity[];
  topPerformers: CampaignRow[];
  executiveSummary: string;
  aiRecommendations: string[];
}