# lib/api-zod/src/generated/types/syncLog.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PerformanceOS AI API
 * OpenAPI spec version: 0.1.0
 */

export interface SyncLog {
  id: string;
  status: string;
  /** @nullable */
  errorMessage?: string | null;
  rowsSynced: number;
  timestamp: string;
}