# lib/api-zod/src/generated/types/windsorConnectionInfo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PerformanceOS AI API
 * OpenAPI spec version: 0.1.0
 */
import type { SyncLog } from './syncLog';

export interface WindsorConnectionInfo {
  connected?: boolean;
  /** @nullable */
  syncStatus?: string | null;
  /** @nullable */
  lastSyncAt?: string | null;
  /** @nullable */
  createdAt?: string | null;
  syncLogs?: SyncLog[];
}